requirements.txt
aiohttp==2.2.5
arrow==0.10.0
asn1crypto==0.22.0
astroid==1.5.3
async-timeout==1.3.0
asyncio==3.4.3
AtomicPuppy==0.4.1
attrs==17.2.0
Automat==0.6.0
behave==1.2.5
binaryornot==0.4.4
certifi==2017.7.27.1
cffi==1.10.0
chardet==3.0.4
click==6.7
constantly==15.1.0
cookiecutter==1.5.1
cryptography==2.0.3
cssselect==1.0.1
defusedxml==0.5.0
Django==1.11.4
django-pipeline==1.6.13
django-social-auth==0.7.28
dpath==1.4.0
et-xmlfile==1.0.1
falcon==1.2.0
future==0.16.0
futures==3.1.1
gevent==1.2.2
greenlet==0.4.12
gunicorn==19.7.1
httplib2==0.10.3
hyperlink==17.3.0
idna==2.6
incremental==17.5.0
isort==4.2.15
jdcal==1.3
Jinja2==2.9.6
jinja2-time==0.2.0
lazy-object-proxy==1.3.1
linkedin-scraping==0.1.2
loremipsum==1.0.5
lxml==3.8.0
MarkupSafe==1.0
mccabe==0.6.1
multidict==3.1.3
oauth2==1.9.0.post1
oauthlib==2.0.2
odfpy==1.3.5
openpyxl==2.4.8
parse==1.8.2
parse-type==0.3.4
parsel==1.2.0
poyo==0.4.1
psycopg2==2.7.3.1
pyasn1==0.3.2
pyasn1-modules==0.0.11
pybreaker==0.4.0
pycparser==2.18
PyDispatcher==2.0.5
PyJWT==1.5.2
pylint==1.7.2
pyOpenSSL==17.2.0
python-dateutil==2.6.1
python-decouple==3.1
python-mimeparse==1.6.0
python-openid==2.2.5
python3-openid==3.1.0
pytz==2017.2
PyYAML==3.12
queuelib==1.4.2
redis==2.10.5
requests==2.18.4
requests-oauthlib==0.8.0
retrying==1.3.3
Scrapy==1.4.0
selenium==3.5.0
service-identity==17.0.0
six==1.10.0
social-auth-app-django==1.2.0
social-auth-core==1.4.0
SQLAlchemy==1.1.13
tablib==0.11.5
Twisted==17.5.0
unicodecsv==0.14.1
Unipath==1.1
urllib3==1.22
virtualenv==15.1.0
w3lib==1.18.0
whichcraft==0.4.1
wrapt==1.10.11
xlrd==1.0.0
xlwt==1.2.0
yarl==0.12.0
zope.interface==4.4.2
